move disconnect to own file
